requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-community>=0.1.0
openai>=1.40.0
anthropic>=0.25.0

# Vector store and embeddings
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Code parsing (Latest versions that support Windows + Python >=3.8,<3.12)
tree-sitter>=0.20.4
tree-sitter-python>=0.25.0
tree-sitter-javascript>=0.25.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.24.3
